+++ libs/argos/src/index.ts
@@ -2,6 +2,9 @@
 export { ArgosModule } from "./argos.module"
 export { type ArgosOptions, ARGOS_OPTIONS } from "./argos.options"
 
+// Interfaces
+export { type AuditableEvent } from "./interfaces/auditable-event.interface"
+
 // Decorators
 export { CreatedBy } from "./decorators/created-by.decorator"
 export { UpdatedBy } from "./decorators/updated-by.decorator"